break(router): move `CronEvent` to `worktop/cfw` module
