+++ packages/worktop/src/sw.d.ts
@@ -1,14 +1,11 @@
-import type { Context, CronEvent, Initializer } from 'worktop';
-import type { Promisable } from 'worktop/utils';
+import type { Context, Initializer } from 'worktop';
 
 declare global {
-	function addEventListener(type: 'fetch', handler: FetchHandler): void;
-	function addEventListener(type: 'scheduled', handler: CronHandler): void;
+	function addEventListener(
+		type: 'fetch',
+		handler: (event: FetchEvent) => void
+	): void;
 }
-
-export type CronHandler = (event: CronEvent) => Promisable<void>;
-export type FetchHandler = (event: FetchEvent) => void;
-
 
 /**
  * Assign the Module `Initializer` as the "fetch" event listener.
